refactored some components
